Made Entity method argument types a little more consistent with the rest of the implementation

+++ src/HTTP/Entity/newData.def.hpp
@@ -1,5 +1,5 @@
 Data::T
-T::newData (size_t entity_size, size_t temp_file_threshhold)
+T::newData (uint64_t entity_size, uint64_t temp_file_threshhold)
 {
 	if (entity_size < temp_file_threshhold)
 	{
